refactor(permission_policy): graduate to top-level layout, add docs-site page
Merge `main` and relocate from `experimental/permission_policy` to top-level
`pydantic_ai_harness/permission_policy/`, matching #347/#354. Drops the
experimental import warning, fixes import paths, moves tests. Adds
`docs/permission-policy.md` (nav.json, index.md, parity gate) and flips the README
Tool access control row to Docs. No behavior change.
